Move accessibility changes for NSSearchField into a subclassed control class.
